backend/requirements.txt
flask==2.3.2
psycopg2-binary==2.9.6
prometheus-client==0.16.0
gunicorn==20.1.0
requests==2.31.0